api/requirements.txt
flask
python-dotenv
waitress
pymongo